requirements.txt
Flask==1.1.2
google-cloud
google-cloud-storage
google-cloud-pubsub
google-cloud-bigquery
